refactor(dust-hive): rename managed services commands to up/down
- Create up.ts and down.ts for managed services (temporal + main session)
- Restore start.ts to resume stopped environments (start SDK watch)
- Restore stop.ts to stop all services in an environment
- Update README.md and CLAUDE.md to reflect new command names

This separates managed services (up/down) from environment lifecycle
(start/stop), avoiding confusion about overloaded command names.
